+++ src/sxapi/cli.py
@@ -1,10 +1,9 @@
 import argparse
-import json
 import sys
 
 from setuptools import setup
 
-from sxapi.publicV2.sensordata import get_sensor_data_for_animal
+from sxapi.cli_sub_parser import create_gsd_parser
 
 
 class cli:
@@ -21,12 +20,12 @@ class cli:
     @staticmethod
     def parse_args(args):
         """Parse arguments from the CLI."""
-        parser = argparse.ArgumentParser(
+        main_parser = argparse.ArgumentParser(
             description=(
                 "Issue calls to the smaXtec system API to import and export data."
             )
         )
-        parser.add_argument(
+        main_parser.add_argument(
             "--version",
             action="store_true",
             default=False,
@@ -34,37 +33,13 @@ class cli:
         )
 
         # gsd_parser
-        subparsers = parser.add_subparsers(help="sub-command help")
-        gsd_parser = subparsers.add_parser(
-            "get_sensor_data",
-            aliases=["gsd"],
-            help="get sensor data from animal(by its ID)",
-        )
-        gsd_parser.set_defaults(func=gsd_sub_function)
-
-        gsd_parser.add_argument("animal_id", help="animal you want get data from")
-        gsd_parser.add_argument(
-            "--metrics",
-            "-m",
-            nargs="*",
-            default=None,
-            help="metrics for sensordata",
-        )
-        gsd_parser.add_argument(
-            "--from_date",
-            "-fd",
-            default=None,
-            nargs=1,
-            help="from_date format: YYYY-MM-DD",
-        )
-        gsd_parser.add_argument(
-            "--to_date", "-td", default=None, nargs=1, help="to_date format: YYYY-MM-DD"
-        )
+        subparsers = main_parser.add_subparsers(help="sub-command help")
+        create_gsd_parser(subparsers)
 
         if not args:
-            parser.print_help()
+            main_parser.print_help()
             return
-        return parser.parse_args(args)
+        return main_parser.parse_args(args)
 
     def run(self):
         """Call sxapi functions based on passed arguments."""
@@ -76,22 +51,6 @@ class cli:
             self.version_info()
         else:
             args.func(args)
-
-
-def gsd_sub_function(args):
-    """
-    function which gets call if the arguments were parsed
-    with get_sensor_data sub-parser
-    """
-    id = args.animal_id
-    metrics = args.metrics
-    from_date = args.from_date
-    to_date = args.to_date
-    resp = get_sensor_data_for_animal(
-        id, metrics=metrics, from_date=from_date, to_date=to_date
-    )
-    if resp is not None:
-        print(json.dumps(resp, indent=0))
 
 
 def cli_run():
